+++ pimonova_WebAPI/Controllers/UserController.cs
@@ -25,15 +25,25 @@ namespace pimonova_WebAPI.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             var Users = await _userRepo.GetAllAsync();
             var UserDTO = Users.Select(s => s.ToUserDTO());
 
             return Ok(Users);
         }
 
-        [HttpGet("{Id}")]
+        [HttpGet("{Id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int Id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             var User = await _userRepo.GetByIdAsync(Id);
 
             if (User == null)
@@ -44,10 +54,15 @@ namespace pimonova_WebAPI.Controllers
             return Ok(User.ToUserDTO());
         }
 
-        [HttpPost("{CompanyId}")]
+        [HttpPost("{CompanyId:int}")]
         public async Task<IActionResult> Create([FromRoute] int CompanyId, CreateUserRequestDTO UserRequestDTO)
         {
-            if(!await _companyRepo.CompanyExists(CompanyId))
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
+            if (!await _companyRepo.CompanyExists(CompanyId))
             {
                 return BadRequest("Company does not exist");
             }
@@ -59,9 +74,14 @@ namespace pimonova_WebAPI.Controllers
         }
 
         [HttpPut]
-        [Route("{Id}")]
+        [Route("{Id:int}")]
         public async Task<IActionResult> Update([FromRoute] int Id, [FromBody] UpdateUserRequestDTO UpdateUserDTO)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             var UserModel = await _userRepo.UpdateAsync(Id, UpdateUserDTO.ToUserFromUpdateDTO());
 
             if (UserModel == null)
@@ -73,9 +93,14 @@ namespace pimonova_WebAPI.Controllers
         }
 
         [HttpDelete]
-        [Route("{Id}")]
+        [Route("{Id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int Id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             var UserModel = await _userRepo.DeleteAsync(Id);
 
             if (UserModel == null)
